Added translation only for bm25
Since bm25 is text based, it is not multilingual. On the opposite,
multilingual-e5-large-instruct is multilingual. Thus, we need to
translate only bm25 text into the language of the DB.

NOTE: Need to rethink how this will work in multilingual setting. Doing
two bm25 searches for each language will probably do it, need to test
time to search.